+++ frontend/src/components/student/student-exams-sections.tsx
@@ -14,7 +14,6 @@ function formatCountdown(seconds: number) {
   if (minutes > 0) return `${minutes}m ${secs}s`
   return `${secs}s`
 }
-
 export function StudentTodayExamsSection({
   examsToday,
   studentClass,
@@ -76,7 +75,6 @@ export function StudentTodayExamsSection({
     </div>
   )
 }
-
 export function StudentUpcomingExamsSection({
   upcomingExams,
   todaysExams,
@@ -120,65 +118,61 @@ export function StudentUpcomingExamsSection({
     </div>
   )
 }
-
 export function StudentCompletedExamsSection({
   results,
 }: {
   results: ExamResult[]
 }) {
   return (
-    <div>
-      <h2 className="text-lg font-semibold mb-3">Completed Exams</h2>
-      {results.length === 0 ? (
-        <Card>
-          <CardContent className="py-6 text-center text-muted-foreground">
-            No completed exams yet
-          </CardContent>
-        </Card>
-      ) : (
-        <div className="space-y-3">
-          {results.map((result) => {
-            const exam = exams.find((entry) => entry.id === result.examId)
-            const percentage = Math.round((result.score / result.totalPoints) * 100)
-            const variant =
-              percentage >= 70 ? 'default' : percentage >= 50 ? 'secondary' : 'destructive'
-            const isReportAvailable = isExamReportAvailable(result.examId)
+    <div><h2 className="text-lg font-semibold mb-3">Completed Exams</h2>{results.length === 0 ? (
+      <Card>
+        <CardContent className="py-6 text-center text-muted-foreground">
+          No completed exams yet
+        </CardContent>
+      </Card>
+    ) : (
+      <div className="space-y-3">
+        {results.map((result) => {
+          const exam = exams.find((entry) => entry.id === result.examId)
+          const percentage = Math.round((result.score / result.totalPoints) * 100)
+          const variant =
+            percentage >= 70 ? 'default' : percentage >= 50 ? 'secondary' : 'destructive'
+          const isReportAvailable = isExamReportAvailable(result.examId)
 
-            return (
-              <Card key={`${result.examId}-${result.studentId}`}>
-                <CardContent className="py-4">
-                  <div className="flex items-center justify-between gap-4">
-                    <div>
-                      <div className="font-medium">{exam?.title}</div>
-                      <div className="text-sm text-muted-foreground">
-                        Submitted: {new Date(result.submittedAt).toLocaleString()}
-                      </div>
-                      <div className="text-sm text-muted-foreground mt-1">
-                        {isReportAvailable
-                          ? 'Detailed report is available'
-                          : 'Detailed report will unlock after every class finishes the exam'}
-                      </div>
+          return (
+            <Card key={`${result.examId}-${result.studentId}`}>
+              <CardContent className="py-4">
+                <div className="flex items-center justify-between gap-4">
+                  <div>
+                    <div className="font-medium">{exam?.title}</div>
+                    <div className="text-sm text-muted-foreground">
+                      Submitted: {new Date(result.submittedAt).toLocaleString()}
                     </div>
-                    <div className="text-right">
-                      <Badge variant={variant}>{percentage}%</Badge>
-                      <div className="text-sm text-muted-foreground mt-1">
-                        {result.score}/{result.totalPoints} points
-                      </div>
-                      <div className="mt-3">
-                        <Link href={`/student/reports/${result.examId}`}>
-                          <Button size="sm" variant="outline">
-                            {isReportAvailable ? 'View Report' : 'Report Locked'}
-                          </Button>
-                        </Link>
-                      </div>
+                    <div className="text-sm text-muted-foreground mt-1">
+                      {isReportAvailable
+                        ? 'Detailed report is available'
+                        : 'Detailed report will unlock after every class finishes the exam'}
                     </div>
                   </div>
-                </CardContent>
-              </Card>
-            )
-          })}
-        </div>
-      )}
-    </div>
+                  <div className="text-right">
+                    <Badge variant={variant}>{percentage}%</Badge>
+                    <div className="text-sm text-muted-foreground mt-1">
+                      {result.score}/{result.totalPoints} points
+                    </div>
+                    <div className="mt-3">
+                      <Link href={`/student/reports/${result.examId}`}>
+                        <Button size="sm" variant="outline">
+                          {isReportAvailable ? 'View Report' : 'Report Locked'}
+                        </Button>
+                      </Link>
+                    </div>
+                  </div>
+                </div>
+              </CardContent>
+            </Card>
+          )
+        })}
+      </div>
+    )}</div>
   )
 }
